Organized js assets to vendor/assets
