+++ app/assets/javascripts/rails_admin/custom/ui.js
@@ -1,17 +1,13 @@
 //= require 'rails_admin/clipboard.min'
-//= require 'rails_admin/disabled-event'
 //= require 'horsey'
 //= require 'rails_admin/owl.carousel'
 //= require rails_admin/utils
-//= require rails_admin/toggle-buttons.js
 //= require rails_admin/triggers-box
 //= require rails_admin/highlight_js/highlight.pack.js
 //= require rails_admin/highcharts
 //= require rails_admin/chartkick
 //= require lodash.min
 //= require rails_admin/select2.full.min
-//= require rails_admin/jquery.easypiechart.min
-//= require rails_admin/king-chart-stat
 //=  require 'rails_admin/custom/cenit'
 
 $(document).on('rails_admin.dom_ready', function () {
