--- app/assets/javascripts/rails_admin/disabled-event.js
@@ -1,10 +1,0 @@
-(function ($) {
-    $.propHooks.disabled = {
-        set: function (el, value) {
-            if (el.disabled !== value) {
-                el.disabled = value;
-                $(el).trigger(value ? 'disabled' : 'enabled');
-            }
-        }
-    };
-})(jQuery);
