+++ app/views/layouts/cenit.html.haml
@@ -6,6 +6,6 @@
     = csrf_meta_tag
     = stylesheet_link_tag "application.css", media: :all
     = stylesheet_link_tag "rails_admin/rails_admin.css", media: :all
-    = javascript_include_tag "rails_admin/rails_admin.js"
   %body
     = yield
+    = javascript_include_tag "rails_admin/rails_admin.js"
